+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## WIP
+
+### Fixed
+
+- Fix Drop component does not immediately show if current active Drop is delaying
+
 ## 3.21.3 (May 22, 2024)
 
 ### Fixed